Add 256-color themes
This change:

* Adds two new themes, "builtin:light256" and "builtin:dark256",
variants of the previous built-in themes but with 256-color palette
defaults for usernames only.

* Adds a new data type, ColorMode = { ColorMode16, ColorMode256 }

* Adds a new InternalTheme field, internalThemeColorMode, indicating the
color palette mode of the theme.

* Adds a new ChatResources field, _crThemeColorMode, indicating the
color palette mode of the active theme.

* Populates the 16- and 256-color themes with default attributes for
username colors and also ensures that 256-color themes can still have
customized colors for each username hash bucket as before, but with a
much larger bucket count.

* Updates Markdown rendering so that instead of always choosing from
the username color list drawn from the 16-color palette as before, we
now require a ColorMode argument (taken from the application state)
to determine how many hash buckets to use when assigning attributes
to usernames. This aspect of the change created the most churn where
ColorMode values are now passed as arguments to quite a few functions.
Essentially this was implicit global state before and is now explicit
application state.